+++ src/features/authentication/confirm-user/model/use-send-one-time-password.test.ts
@@ -1,10 +1,10 @@
 import router from '@app/providers/router'
-import { getCurrentSession } from '@features/authentication/model/get-current-session'
+import { getCurrentSession } from '@features/authentication/service/get-current-session'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { monitorInformation, MONITORING_EVENTS } from '../../monitoring/monitoring'
-import { getRegisterEmailFromStorage } from '../register-storage'
-import { checkOneTimePassword } from './check-one-time-password'
+import { monitorInformation, MONITORING_EVENTS } from '../monitoring/monitoring'
+import { checkOneTimePassword } from '../service/check-one-time-password'
+import { getRegisterEmailFromStorage } from './register-storage'
 import { useSendOneTimePassword, verifyOtp } from './use-send-one-time-password'
 
 let onMountedHandler: (() => Promise<void>) | undefined
@@ -26,19 +26,19 @@ vi.mock('@app/providers/router', () => ({
   }
 }))
 
-vi.mock('@features/authentication/model/get-current-session', () => ({
+vi.mock('@features/authentication/service/get-current-session', () => ({
   getCurrentSession: vi.fn()
 }))
 
-vi.mock('./check-one-time-password', () => ({
+vi.mock('../service/check-one-time-password', () => ({
   checkOneTimePassword: vi.fn()
 }))
 
-vi.mock('../register-storage', () => ({
+vi.mock('./register-storage', () => ({
   getRegisterEmailFromStorage: vi.fn()
 }))
 
-vi.mock('../../monitoring/monitoring', () => ({
+vi.mock('../monitoring/monitoring', () => ({
   MONITORING_EVENTS: {
     CHECK_OTP: 'CHECK_OTP'
   },
