+++ src/features/authentication/confirm-user/service/resend-sign-up-confirmation.ts
