+++ src/features/authentication/register-user/model/use-register.ts
@@ -2,8 +2,8 @@ import { ref } from 'vue'
 import type { AuthActionResult } from '@shared/types'
 
 import { monitorInformation, MONITORING_EVENTS } from '../monitoring/monitoring'
-import { setIsOtpActiveToStorage, setRegisterEmailToStorage } from './register-storage'
-import { signUpWithMailAndPassword } from './sign-up-with-mail-and-password'
+import { setIsOtpActiveToStorage, setRegisterEmailToStorage } from '../service/register-storage'
+import { signUpWithMailAndPassword } from '../service/sign-up-with-mail-and-password'
 
 /**
  * Composable for registering a new user account.
