+++ src/features/authentication/register-user/service/sign-up-with-mail-and-password.test.ts
