+++ src/features/authentication/service/on-auth-state-change.ts
