+++ src/features/authentication/service/use-auth-navigation.ts
