+++ src/features/authentication/signin-user/model/use-login.test.ts
@@ -1,10 +1,10 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
 import { monitorInformation, MONITORING_EVENTS } from '../monitoring/monitoring'
-import { signInWithEmailPassword } from './sign-in-with-email-password'
+import { signInWithEmailPassword } from '../service/sign-in-with-email-password'
 import { loginUserAccount, useLogin } from './use-login'
 
-vi.mock('./sign-in-with-email-password', () => ({
+vi.mock('../service/sign-in-with-email-password', () => ({
   signInWithEmailPassword: vi.fn()
 }))
