+++ src/features/authentication/signin-user/model/use-login.ts
@@ -2,7 +2,7 @@ import { ref } from 'vue'
 import type { AuthActionResult } from '@shared/types'
 
 import { monitorInformation, MONITORING_EVENTS } from '../monitoring/monitoring'
-import { signInWithEmailPassword } from './sign-in-with-email-password'
+import { signInWithEmailPassword } from '../service/sign-in-with-email-password'
 
 /**
  * Runs sign-in via {@link loginUserAccount}.
